backend/requirements.txt
attrs==23.1.0
blinker==1.6.2
CacheControl==0.12.11
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
filelock==3.12.0
firebase-admin==6.1.0
Flask==2.3.2
gcloud==0.17.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.18.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.54.2
grpcio-status==1.54.2
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jws==0.1.3
llvmlite==0.40.0
MarkupSafe==2.1.2
more-itertools==9.1.0
mpmath==1.3.0
msgpack==1.0.5
networkx==3.1
numba==0.57.0
numpy==1.24.3
oauth2client==3.0.0
openai-whisper @ git+https://github.com/openai/whisper.git@248b6cb124225dd263bb9bd32d060b6517e067f8
pandas==2.0.1
proto-plus==1.22.2
protobuf==4.23.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.4.3
pydub==0.25.1
pygame==2.4.0
PyJWT==2.7.0
pyparsing==3.0.9
pyproject-toml==0.0.10
Pyrebase==3.0.27
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jwt==2.0.1
pytz==2023.3
regex==2023.5.5
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sympy==1.12
threadpoolctl==3.1.0
tiktoken==0.3.3
toml==0.10.2
torch==2.0.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
Werkzeug==2.3.4
